Move FontsDataSet to a separate file
