+++ foo_ui_columns/foo_ui_columns.vcxproj.filters
@@ -548,6 +548,9 @@
     <ClCompile Include="fcl_colours.cpp">
       <Filter>FCL files</Filter>
     </ClCompile>
+    <ClCompile Include="fcl_fonts.cpp">
+      <Filter>FCL files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include=".\resource.h">
